Refactored package streeplijst to streeplijst2. Renamed items.py to streeplijst.py
Refactored all other files accordingly.
