+++ requirements.txt
@@ -1,0 +1,3 @@
+setuptools~=49.2.1
+requests~=2.24.0
+Flask~=1.1.2
